[move-text-list-editor] Move InfoTip
