+++ packages/perseus-editor/src/components/info-tip/index.tsx
