+++ packages/perseus-editor/src/components/info-tip/info-tip-base.tsx
