+++ packages/perseus-editor/src/item-extras-editor.tsx
@@ -1,4 +1,3 @@
-import {components} from "@khanacademy/perseus";
 import {
     ItemExtras,
     getDefaultAnswerArea,
@@ -9,6 +8,8 @@ import {Checkbox} from "@khanacademy/wonder-blocks-form";
 import {spacing} from "@khanacademy/wonder-blocks-tokens";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
+
+import InfoTip from "./components/info-tip";
 
 import type {APIOptions} from "@khanacademy/perseus";
 import type {
@@ -37,8 +38,6 @@ const calculatorVariants: Array<{
         variant: "four_function",
     },
 ];
-
-const {InfoTip} = components;
 
 type Props = PerseusAnswerArea & {
     apiOptions?: APIOptions;
