+++ packages/perseus-editor/src/widgets/interactive-graph-editor/components/polygon-answer-options.tsx
@@ -1,10 +1,10 @@
-import {components} from "@khanacademy/perseus";
 import {View} from "@khanacademy/wonder-blocks-core";
 import {Checkbox} from "@khanacademy/wonder-blocks-form";
 import {BodyText} from "@khanacademy/wonder-blocks-typography";
 import * as React from "react";
 import invariant from "tiny-invariant";
 
+import InfoTip from "../../../components/info-tip";
 import {TypedSingleSelect} from "../../../components/typed-single-select";
 import {parsePointCount} from "../../../util/points";
 import styles from "../interactive-graph-editor.module.css";
@@ -15,8 +15,6 @@ import type {
     PerseusGraphType,
     PerseusGraphTypePolygon,
 } from "@khanacademy/perseus-core";
-
-const {InfoTip} = components;
 
 const POLYGON_SIDES_OPTIONS: Record<string, string> = {
     "3": "3 sides",
