+++ packages/perseus-editor/src/widgets/orderer-editor.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable @khanacademy/ts-no-error-suppressions */
 /* eslint-disable react/forbid-prop-types */
-import {components} from "@khanacademy/perseus";
 import {
     ordererLogic,
     type OrdererDefaultWidgetOptions,
@@ -8,9 +7,8 @@ import {
 import PropTypes from "prop-types";
 import * as React from "react";
 
+import InfoTip from "../components/info-tip";
 import TextListEditor from "../components/text-list-editor";
-
-const {InfoTip} = components;
 
 const NORMAL = "normal";
 const AUTO = "auto";
