+++ packages/perseus/src/components.ts
@@ -2,7 +2,6 @@ export {default as ButtonGroup} from "./components/button-group";
 export {default as Graphie} from "./components/graphie";
 export {default as HUD} from "./components/hud";
 export {default as Icon} from "./components/icon";
-export {default as InfoTip} from "./components/info-tip";
 export {default as InlineIcon} from "./components/inline-icon";
 export {default as MathInput} from "./components/math-input";
 export {default as MultiButtonGroup} from "./components/multi-button-group";
